Moving src files to separate dirs
